filmsite/requirements.txt
Django==3.2.2
django-model-utils==4.1.1
djangorestframework==3.12.4
Pillow==7.0.0
